refactoring code, moving env related methods to a new module
